+++ src/lib.rs
@@ -1,5 +1,5 @@
 #![cfg_attr(not(any(test, feature = "std")), no_std)]
-#![doc(html_root_url = "https://docs.rs/geometrid/0.7.0")]
+#![doc(html_root_url = "https://docs.rs/geometrid/0.8.0")]
 // #![deny(missing_docs)]
 #![deny(warnings, dead_code, unused_imports, unused_mut)]
 #![warn(clippy::pedantic)]